requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-BasicAuth==0.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jsonlines==1.2.0
MarkupSafe==1.1.1
oauthlib==3.1.0
PySocks==1.7.1
python-dotenv==0.14.0
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
tweepy==3.9.0
urllib3==1.25.11
Werkzeug==1.0.1
